DR 110072: Separate batched components (part 4)
